Move private_ip_addresses
